requirements.txt
TikTokApi>=6.5.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
playwright>=1.40.0
apscheduler>=3.10.0
pytrends>=4.9.0

# Live trading (optional)
solana>=0.34.0
solders>=0.21.0
base58>=2.1.0
